Update README.md
Updated / Re-wrote the NAT explanation to use NAPT (Network Address Port Translation). Also added in DHCP as that is what creates the local network and gives devices local IP addresses, I think the way it was worded before "It does this by creating a local network, providing local IP addresses to all machines that connect" was wrong which is why i put in this pull request. P.s im a big fan of your work 👍